🎨  refactor: Refactor button component and update imports in Home and Profile screens
